fix bug in stutter filter
As currently written the stutter removal will only be applied to the
first stutter entry.  If the next-highest-count entry below that also
looks like stutter, it won't be filtered.  This changes the logic to
remove as many in a row as needed so the second called allele is never
an apparent stutter sequence.