Make report_genotypes handle length homoplasy
This appends an asterisk to the name of the second allele for casees
where the two alleles are labeled with the same name but have different
sequence content.  This should help highlight cases of length homplasy
in particular where names are only using sequence lengths.  (This was
already done in the heatmaps; the logic should be consolidated at some
point.)