cleanup lingering Locus factor in full_analysis
I had been using a factor of locus identifiers in full_analysis to
enforce the same ordering as in locus_attrs, but that leaves it as a
factor downstream.  This instead uses match() and two order() calls to
order primarily by locus as in locus_attrs and then by all other columns
as defined in order_entries().