rewrite find_stutter and find_artifact
These previously had nested for loops that became very inefficient for
large numbers of unique sequences.  Now the inner loops are replaced
with vector operations which I think should speed it up considerably in
these cases.  (The outer per-locus loop remains but it doesn't have a
significant effect on the timing anyway.)