Switch to compiled app for Mac OS drag-and-drop
Apparently the AppleScript must be compiled to binary and bundled into a
.app directory within Mac OS for a drag-and-drop to work as expected.
This puts a .app into the inst/bin directory and updates scripts
accordingly.  There are some binaries in there, but the whole tree is
quite small, so keeping within git seems OK as it shouldn't be changing
often.