move installer code into package proper
These functions may be helpful from within an install (e.g. to keep the
Desktop icon valid as #24 describes).  This also removes the brittle
attempts to autodetect paths and instead requires that the wrapper
install script be called from the project dir and install() function be
given a package source path.